Make chain/network/src/routing pub(crate)
