+++ chain/network/Cargo.toml
@@ -57,9 +57,5 @@ name = "graph"
 harness = false
 
 [[bench]]
-name = "ibf"
-harness = false
-
-[[bench]]
 name = "routing_table_actor"
 harness = false
