+++ chain/network/src/lib.rs
@@ -1,5 +1,4 @@
 pub use near_network_primitives::types::PeerInfo;
-
 pub use peer_manager::peer_manager_actor::PeerManagerActor;
 pub use peer_manager::peer_store::iter_peers_from_store;
 pub use routing::routing_table_actor::{
