+++ chain/network/src/routing/mod.rs
@@ -1,10 +1,13 @@
 pub(crate) mod codec;
-pub mod edge;
+pub(crate) mod edge;
 pub(crate) mod edge_verifier_actor;
+#[cfg(feature = "protocol_feature_routing_exchange_algorithm")]
 pub(crate) mod ibf;
-pub mod ibf_peer_set;
-pub mod ibf_set;
+#[cfg(feature = "protocol_feature_routing_exchange_algorithm")]
+pub(crate) mod ibf_peer_set;
+#[cfg(feature = "protocol_feature_routing_exchange_algorithm")]
+pub(crate) mod ibf_set;
 mod route_back_cache;
-pub mod routing;
-pub mod routing_table_actor;
+pub(crate) mod routing;
+pub(crate) mod routing_table_actor;
 mod utils;
