+++ chain/network/tests/cache.rs
@@ -1,5 +1,4 @@
 use near_crypto::Signature;
-use near_network::routing::routing::RoutingTableView;
 use near_network::test_utils::{random_epoch_id, random_peer_id};
 use near_primitives::network::AnnounceAccount;
 use near_store::test_utils::create_test_store;
