+++ integration-tests/tests/network/infinite_loop.rs
@@ -10,7 +10,7 @@ use near_actix_test_utils::run_actix;
 use near_client::ClientActor;
 use near_logger_utils::init_integration_logger;
 
-use near_network::routing::routing_table_actor::start_routing_table_actor;
+use near_network::routing::start_routing_table_actor;
 use near_network::test_utils::{convert_boot_nodes, open_port, GetInfo, WaitOrTimeoutActor};
 use near_network::types::{PeerManagerMessageRequest, SyncData};
 use near_network::{NetworkClientResponses, NetworkRequests, PeerManagerActor};
