+++ integration-tests/tests/network/peer_handshake.rs
@@ -13,7 +13,7 @@ use near_actix_test_utils::run_actix;
 use near_client::{ClientActor, ViewClientActor};
 use near_logger_utils::init_test_logger;
 
-use near_network::routing::routing_table_actor::start_routing_table_actor;
+use near_network::routing::start_routing_table_actor;
 
 use near_network::test_utils::{
     convert_boot_nodes, open_port, GetInfo, StopSignal, WaitOrTimeoutActor,
